Fix local variables allocation
When function is called it is lazily parsed in order to
deduct arguments types. It also creates a nested scope (context).
Scopes have start and last addresses for variables allocations, and
the allocation is based on parent's scope last used address.

Because the child scope starts growing while parent was not completed
this can cause problems with when the same function is called again later
and the parent allocated some vars between invocations.

In order to fix this scopes gets remapped on each function call
up to the leaves by updating.
Additionally, regular functions now pass arguments via stack.